End of SgnUpController setting

+++ src/main/java/app/foodapp/controller/login/SignUpController.java
@@ -1,8 +1,10 @@
 package app.foodapp.controller.login;
 
+import app.foodapp.view.alert.AlertFound;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
@@ -41,13 +43,36 @@ public class SignUpController {
     }
 
     public void userRegister() throws IOException {
-        String content = username.getText() + "," + password.getText() + "," + regime;
-        File file = new File("userInformation.txt");
-        if(!file.exists())
-            file.createNewFile();
-        FileWriter fw = new FileWriter(file.getAbsoluteFile(), true);
-        fw.append(content);
-        fw.close();
-
+        String content = username.getText() + "," + password.getText() + "," + regime + ",\n";
+        if(userAlreadyExist(content))
+            AlertFound.usernameAlreadyExist();
+        if(username.getText().isEmpty())
+            AlertFound.usernameNotWrite();
+        if(password.getText().isEmpty())
+            AlertFound.passwordNotWrite();
+        if(!userAlreadyExist(content)){
+            File file = new File("userInformation.txt");
+            if(!file.exists())
+                file.createNewFile();
+            FileWriter fw = new FileWriter(file.getAbsoluteFile(), true);
+            fw.append(content);
+            fw.close();
+        }
     }
+
+    private Boolean userAlreadyExist(String content) throws IOException {
+        Boolean usernameAlreadyExist = false;
+        File userInformation = new File("userInformation.txt");
+        BufferedReader obj = new BufferedReader(new FileReader(userInformation));
+        String strng;
+        while ((strng = obj.readLine()) != null) {
+            strng = strng + "\n";
+            if (strng.equals(content)) {
+                usernameAlreadyExist = true;
+            }
+        }
+
+        return usernameAlreadyExist;
+    }
+
 }
